Delete old implementation, non-used key not found and move skipXXX
