+++ aws/logs_monitoring_go/internal/parsing/bench_test.go
@@ -41,86 +41,44 @@ var (
 	benchEventBridge = json.RawMessage(`{"version":"0","detail-type":"Object Created","source":"aws.s3","detail":{"bucket":{"name":"my-bucket"},"object":{"key":"my-key"}}}`)
 )
 
-func BenchmarkParse_Decoder_CloudWatch(b *testing.B) {
+func BenchmarkParse_CloudWatch(b *testing.B) {
 	b.ReportAllocs()
 	for b.Loop() {
 		_, _ = Parse(benchCloudWatch)
 	}
 }
 
-func BenchmarkParse_Unmarshal_CloudWatch(b *testing.B) {
-	b.ReportAllocs()
-	for b.Loop() {
-		_, _ = ParseUnmarshal(benchCloudWatch)
-	}
-}
-
-func BenchmarkParse_Decoder_S3(b *testing.B) {
+func BenchmarkParse_S3(b *testing.B) {
 	b.ReportAllocs()
 	for b.Loop() {
 		_, _ = Parse(benchS3)
 	}
 }
 
-func BenchmarkParse_Unmarshal_S3(b *testing.B) {
-	b.ReportAllocs()
-	for b.Loop() {
-		_, _ = ParseUnmarshal(benchS3)
-	}
-}
-
-func BenchmarkParse_Decoder_SNS_S3(b *testing.B) {
+func BenchmarkParse_SNS_S3(b *testing.B) {
 	b.ReportAllocs()
 	for b.Loop() {
 		_, _ = Parse(benchSNSS3)
 	}
 }
 
-func BenchmarkParse_Unmarshal_SNS_S3(b *testing.B) {
-	b.ReportAllocs()
-	for b.Loop() {
-		_, _ = ParseUnmarshal(benchSNSS3)
-	}
-}
-
-func BenchmarkParse_Decoder_SQS_S3(b *testing.B) {
+func BenchmarkParse_SQS_S3(b *testing.B) {
 	b.ReportAllocs()
 	for b.Loop() {
 		_, _ = Parse(benchSQSS3)
 	}
 }
 
-func BenchmarkParse_Unmarshal_SQS_S3(b *testing.B) {
-	b.ReportAllocs()
-	for b.Loop() {
-		_, _ = ParseUnmarshal(benchSQSS3)
-	}
-}
-
-func BenchmarkParse_Decoder_SQS_SNS_S3(b *testing.B) {
+func BenchmarkParse_SQS_SNS_S3(b *testing.B) {
 	b.ReportAllocs()
 	for b.Loop() {
 		_, _ = Parse(benchSQSSNSS3)
 	}
 }
 
-func BenchmarkParse_Unmarshal_SQS_SNS_S3(b *testing.B) {
-	b.ReportAllocs()
-	for b.Loop() {
-		_, _ = ParseUnmarshal(benchSQSSNSS3)
-	}
-}
-
-func BenchmarkParse_Decoder_EventBridge(b *testing.B) {
+func BenchmarkParse_EventBridge(b *testing.B) {
 	b.ReportAllocs()
 	for b.Loop() {
 		_, _ = Parse(benchEventBridge)
-	}
-}
-
-func BenchmarkParse_Unmarshal_EventBridge(b *testing.B) {
-	b.ReportAllocs()
-	for b.Loop() {
-		_, _ = ParseUnmarshal(benchEventBridge)
 	}
 }
