+++ aws/logs_monitoring_go/internal/parsing/eventbridge.go
@@ -6,7 +6,6 @@
 package parsing
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"strings"
@@ -14,121 +13,49 @@ import (
 	"github.com/aws/aws-lambda-go/events"
 )
 
-const (
-	eventBridgeSourceS3     = "aws.s3"
-	eventBridgeDetailTypeS3 = "Object Created"
-)
+type s3EventBridgeDetail struct {
+	Bucket struct {
+		Name string `json:"name"`
+	} `json:"bucket"`
+	Object struct {
+		Key string `json:"key"`
+	} `json:"object"`
+}
 
-func parseEventBridge(event json.RawMessage) ([]Event, error) {
-	source, detailType, detail, err := decodeEventBridgeEnvelope(event)
-	if err != nil {
-		return nil, fmt.Errorf("decode eventbridge: %w", err)
+func eventBridge(event json.RawMessage) ([]Event, error) {
+	var eventBridgeEvent events.EventBridgeEvent
+	if err := json.Unmarshal(event, &eventBridgeEvent); err != nil {
+		return nil, fmt.Errorf("unmarshal: %w", err)
 	}
 
-	if source == eventBridgeSourceS3 && strings.Contains(detailType, eventBridgeDetailTypeS3) {
-		s3Event, err := buildS3EventFromEventBridge(detail)
-		if err != nil {
-			return nil, fmt.Errorf("build s3 event from eventbridge: %w", err)
+	if eventBridgeEvent.Source == "aws.s3" && strings.Contains(eventBridgeEvent.DetailType, "Object Created") {
+		var s3eb s3EventBridgeDetail
+		if err := json.Unmarshal(eventBridgeEvent.Detail, &s3eb); err != nil {
+			return nil, fmt.Errorf("unmarshal: %w", err)
 		}
-		return []Event{{ContentType: ContentTypeS3, Payload: s3Event}}, nil
-	}
 
+		payload, err := mapS3EventBridge(s3eb)
+		if err != nil {
+			return nil, err
+		}
+
+		return []Event{{ContentType: ContentTypeS3, Payload: payload}}, nil
+	}
 	return []Event{{ContentType: ContentTypeEventBridge, Payload: event}}, nil
 }
 
-func decodeEventBridgeEnvelope(event json.RawMessage) (source, detailType string, detail json.RawMessage, err error) {
-	dec := json.NewDecoder(bytes.NewReader(event))
-	if err := SkipBrace(dec); err != nil {
-		return "", "", nil, err
+func mapS3EventBridge(eb s3EventBridgeDetail) (json.RawMessage, error) {
+	s3EventRecord := events.S3EventRecord{
+		EventSource: eventSourceS3,
+		S3: events.S3Entity{
+			Bucket: events.S3Bucket{Name: eb.Bucket.Name},
+			Object: events.S3Object{Key: eb.Object.Key, URLDecodedKey: eb.Object.Key},
+		},
 	}
-
-	for dec.More() {
-		key, err := dec.Token()
-		if err != nil {
-			return "", "", nil, fmt.Errorf("read key: %w", err)
-		}
-
-		switch key {
-		case "source":
-			if err := dec.Decode(&source); err != nil {
-				return "", "", nil, fmt.Errorf("source: %w", err)
-			}
-		case "detail-type":
-			if err := dec.Decode(&detailType); err != nil {
-				return "", "", nil, fmt.Errorf("detail-type: %w", err)
-			}
-		case "detail":
-			if err := dec.Decode(&detail); err != nil {
-				return "", "", nil, fmt.Errorf("detail: %w", err)
-			}
-		default:
-			if err := Skip(dec); err != nil {
-				return "", "", nil, err
-			}
-		}
-	}
-
-	return source, detailType, detail, nil
-}
-
-func buildS3EventFromEventBridge(detail json.RawMessage) (json.RawMessage, error) {
-	bucketName, objectKey, err := decodeEventBridgeS3Detail(detail)
-	if err != nil {
-		return nil, fmt.Errorf("decode: %w", err)
-	}
-
-	s3Event := events.S3Event{
-		Records: []events.S3EventRecord{{
-			EventSource: eventSourceS3,
-			S3: events.S3Entity{
-				Bucket: events.S3Bucket{Name: bucketName},
-				Object: events.S3Object{Key: objectKey, URLDecodedKey: objectKey},
-			},
-		}},
-	}
-	payload, err := json.Marshal(s3Event)
+	payload, err := json.Marshal(s3EventRecord)
 	if err != nil {
 		return nil, fmt.Errorf("marshal: %w", err)
 	}
+
 	return payload, nil
-}
-
-func decodeEventBridgeS3Detail(detail json.RawMessage) (bucket, key string, err error) {
-	dec := json.NewDecoder(bytes.NewReader(detail))
-
-	if err := SkipBrace(dec); err != nil {
-		return "", "", err
-	}
-
-	for dec.More() {
-		k, err := dec.Token()
-		if err != nil {
-			return "", "", fmt.Errorf("read key: %w", err)
-		}
-
-		switch k {
-		case "bucket":
-			var b struct {
-				Name string `json:"name"`
-			}
-			if err := dec.Decode(&b); err != nil {
-				return "", "", fmt.Errorf("bucket: %w", err)
-			}
-			bucket = b.Name
-		case "object":
-			var o struct {
-				Key string `json:"key"`
-			}
-			if err := dec.Decode(&o); err != nil {
-				return "", "", fmt.Errorf("object: %w", err)
-			}
-			key = o.Key
-		default:
-			if err := Skip(dec); err != nil {
-				return "", "", err
-			}
-		}
-	}
-
-	return bucket, key, nil
 }
